pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "autoreason"
version = "0.1.0"
description = "Iterative multi-agent refinement loop for subjective work"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "John-Alan Simmons" }]
keywords = ["llm", "agents", "refinement", "iteration", "critique"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "litellm>=1.40",
    "pyyaml>=6.0",
    "click>=8.1",
    "rich>=13.0",
    "pydantic>=2.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
autoreason = "autoreason.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/autoreason"]

[tool.hatch.build.targets.wheel.force-include]
"src/autoreason/default_prompts.yaml" = "autoreason/default_prompts.yaml"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
